update move unit tests for module_stats to the same file as prod code, update README.md
